requirements.txt
django
suds-py3